backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.0.0
anthropic>=0.40.0
openai>=1.50.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
structlog>=24.0.0
python-slugify>=8.0.0
